Update SlugShotgunAccessory.txt
Hi I'm relatively new to GitHub so I don't know if this is the proper way to comment on code but I noticed a bug with your mod.
When I found a splatgun with this accessory on it it just wouldn't shoot. Just letting you know, in case you could fix it. I absolutely love your mods, I have many downloaded already.